refactor useAnimatedScan
